requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
reportlab
openpyxl
python-dateutil
aiofiles
xlwt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pytz
